+++ eds-aic-20230930/alibabacloud_eds_aic20230930/models/_describe_cloud_phone_nodes_request.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from typing import List
 
+from alibabacloud_eds_aic20230930 import models as main_models
 from darabonba.model import DaraModel
 
 class DescribeCloudPhoneNodesRequest(DaraModel):
@@ -19,6 +20,7 @@ class DescribeCloudPhoneNodesRequest(DaraModel):
         node_name_list: List[str] = None,
         server_type: str = None,
         status: str = None,
+        tags: List[main_models.DescribeCloudPhoneNodesRequestTags] = None,
     ):
         self.bandwidth_package_id = bandwidth_package_id
         # The region ID.
@@ -51,9 +53,13 @@ class DescribeCloudPhoneNodesRequest(DaraModel):
         # *   DELETED: The matrix is deleted.
         # *   CREATING: The matrix is being created.
         self.status = status
+        self.tags = tags
 
     def validate(self):
-        pass
+        if self.tags:
+            for v1 in self.tags:
+                 if v1:
+                    v1.validate()
 
     def to_map(self):
         result = dict()
@@ -90,6 +96,11 @@ class DescribeCloudPhoneNodesRequest(DaraModel):
         if self.status is not None:
             result['Status'] = self.status
 
+        result['Tags'] = []
+        if self.tags is not None:
+            for k1 in self.tags:
+                result['Tags'].append(k1.to_map() if k1 else None)
+
         return result
 
     def from_map(self, m: dict = None):
@@ -123,6 +134,47 @@ class DescribeCloudPhoneNodesRequest(DaraModel):
 
         if m.get('Status') is not None:
             self.status = m.get('Status')
+
+        self.tags = []
+        if m.get('Tags') is not None:
+            for k1 in m.get('Tags'):
+                temp_model = main_models.DescribeCloudPhoneNodesRequestTags()
+                self.tags.append(temp_model.from_map(k1))
+
+        return self
+
+class DescribeCloudPhoneNodesRequestTags(DaraModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+    ):
+        self.key = key
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.key is not None:
+            result['Key'] = self.key
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
 
         return self